+++ CHEOperations.cpp
@@ -27,18 +27,18 @@ CHEOperations::CHEOperations(const CHEOperations &orig)
 std::vector<unsigned int> CHEOperations::geNeighbourVertices(const unsigned int halfEdge) const
 {
     std::vector<unsigned int> neighbourVertices;
-    long int currentHalfEdge = halfEdge;
+    unsigned int currentHalfEdge = halfEdge;
 
     //Turn around the vertex until get the starting point or hit a boundary.
     do
     {
-        // Get the vertex neighbour to the current half edge.
+        // Get the vertex neighbor to the current half-edge.
         const unsigned int vertex = _che->getHalfEdgeVertexIndex(_che->halfEdgeNext(currentHalfEdge));
 
-        // Save it to the neighbours list.
+        // Save it to the neighbor list.
         neighbourVertices.push_back(vertex);
 
-        // Go the next face.
+        // Go to the next face.
         currentHalfEdge = _che->halfEdgeOpposite(currentHalfEdge);
 
         // If it hits a border, stop.
@@ -47,7 +47,7 @@ std::vector<unsigned int> CHEOperations::geNeighbourVertices(const unsigned int 
             break;
         }
 
-        // Get the next half edge. Basically it goes back to a half edge emanating from the vertex.
+        // Get the next half-edge. Basically it goes back to a half-edge emanating from the vertex.
         currentHalfEdge = _che->halfEdgeNext(currentHalfEdge);
     }
     while (currentHalfEdge != halfEdge);
@@ -56,20 +56,20 @@ std::vector<unsigned int> CHEOperations::geNeighbourVertices(const unsigned int 
     if (currentHalfEdge != -1)
         return neighbourVertices;
 
-    // The loop hit a border. Start from beginning and turn around the vertex in the other orientation.
+    // The loop hit a border. Start from the beginning and turn around the vertex in the other orientation.
     currentHalfEdge = halfEdge;
     do
     {
         // Loop in the opposite orientation.
-        const long int halfEdgePrevious = _che->halfEdgePrevious(currentHalfEdge);
+        const unsigned int halfEdgePrevious = _che->halfEdgePrevious(currentHalfEdge);
 
-        // Get the vertex neighbour to the current half edge.
-        const int vertex = _che->getHalfEdgeVertexIndex(halfEdgePrevious);
+        // Get the vertex neighbor to the current half-edge.
+        const unsigned int vertex = _che->getHalfEdgeVertexIndex(halfEdgePrevious);
 
-        // Save it to the neighbours list.
+        // Save it to the neighbor list.
         neighbourVertices.push_back(vertex);
 
-        // Go the next face.
+        // Go to the next face.
         currentHalfEdge = _che->halfEdgeOpposite(halfEdgePrevious);
     }
     while (currentHalfEdge != -1);
@@ -87,13 +87,13 @@ std::vector<unsigned> CHEOperations::getNeighbourFaces(const unsigned int face) 
     //Get the face's first half edge id.
     const unsigned int halfEdgeInFace = face * numberCoordinatesByVertex;
 
-    // Initialize the current half edge to start looping the face.
+    // Initialize the current half-edge to start looping the face.
     unsigned int currentHalfEdge = halfEdgeInFace;
 
-    // Loop the faces getting the opposite half edge for each edge in order to get the neighbour faces.
+    // Loop the faces getting the opposite half-edge for each edge to get the neighbor faces.
     do
     {
-        // Get the neighbour's face half edge.
+        // Get the neighbor's face half edge.
         const unsigned int opposite = _che->halfEdgeOpposite(currentHalfEdge);
 
         // Check if it is not a border.
@@ -106,7 +106,7 @@ std::vector<unsigned> CHEOperations::getNeighbourFaces(const unsigned int face) 
             neighbourFaces.push_back(faceFound);
         }
 
-        // Go the next face's half edge.
+        // Go the next face's half-edge.
         currentHalfEdge = _che->halfEdgeNext(currentHalfEdge);
     }
     while (currentHalfEdge != halfEdgeInFace);
